# [xy] Speed up analysis calculator