# Summary
<!-- Brief summary of what your code does -->
Speed up analysis calculator. Tested on a dataset with 100k rows and 32 columns (with datetime column). Original time was 2.1s. This PR reduced the analysis calculation time by 66% to 0.7s.
* Speed up timestamp calculation: 2.1s -> 1.5s
  *  https://stackoverflow.com/questions/54313463/pandas-datetime-to-unix-timestamp-seconds
* Optimize stats calculation in timeseries chart: 1.5s -> 1.15s 
* Update histogram calculation to use np.histogram method: 1.15s -> 0.78s
* Only dropna when necessary for histogram calculation: 0.78s -> 0.7s

# Tests
<!-- How did you test your change? -->
tested locally

cc:
<!-- Optionally mention someone to let them know about this pull request -->